refactor: MemberService만 MemberDAO랑 소통하도록 통합
