+++ server/src/main/java/com/emr/slgi/member/controller/StaffController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.emr.slgi.member.service.StaffService;
+import com.emr.slgi.member.service.MemberService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -16,10 +16,10 @@ import lombok.RequiredArgsConstructor;
 @RequestMapping("/staff")
 @RequiredArgsConstructor
 public class StaffController {
-  private final StaffService staffService;
+  private final MemberService memberService;
 
   @GetMapping("/list/{uuid}")
   public ResponseEntity<?> getStaffList(@PathVariable("uuid") String uuid) {
-    return ResponseEntity.ok(Map.of("list", staffService.getStaffList(uuid)));
+    return ResponseEntity.ok(Map.of("list", memberService.getStaffList(uuid)));
   }
 }
